server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
pyjwt==2.9.0
bcrypt==4.2.0
google-auth==2.35.0
python-dotenv==1.0.1
alembic==1.13.0
pydantic[email]==2.9.0
requests==2.32.0
